Roll test

Tests for roll module.

See merge request !50

+++ src/dungeonbot/models/roll.py
@@ -1,4 +1,5 @@
 from sqlalchemy.orm.exc import NoResultFound
+from sqlalchemy.exc import IntegrityError
 from dungeonbot.models import db
 from datetime import datetime
 
@@ -12,7 +13,7 @@ class RollModel(db.Model):
     """
 
     id = db.Column(db.Integer, primary_key=True)
-    key = db.Column(db.String(256))
+    key = db.Column(db.String(256), unique=True)
     val = db.Column(db.String(256))
     user = db.Column(db.String(256))
     created = db.Column(db.DateTime, nullable=False, default=datetime.utcnow())
@@ -22,18 +23,24 @@ class RollModel(db.Model):
         """Create New saved roll from a key value pair."""
         if session is None:
             session = db.session
-        instance = cls(key=key, val=val, user=user)
-        session.add(instance)
-        session.commit()
-        return instance
+        if not key or not val:
+            return None
+        try:
+            instance = cls(key=key, val=val, user=user["id"])
+            session.add(instance)
+            session.commit()
+            return instance
+        except IntegrityError:
+            session.rollback()
+            return "duplicate"
 
     @classmethod
-    def get_by_key(cls, key=None, user=None, session=None):
+    def get(cls, key=None, user=None, session=None):
         """Query Database by name (Key)."""
         if session is None:
             session = db.session
         try:
-            instance = session.query(cls).filter_by(key=key, user=user).first()
+            instance = session.query(cls).filter_by(key=key, user=user["id"]).one()
         except NoResultFound:
             instance = None
         return instance
@@ -43,10 +50,12 @@ class RollModel(db.Model):
         """Delete Roll."""
         if session is None:
             session = db.session
+        if not isinstance(instance, RollModel):
+            return None
         name = instance.key
         session.delete(instance)
         session.commit()
-        return name + " was successfully deleted."
+        return name
 
     @classmethod
     def list(cls, how_many=10, user=None, session=None):
@@ -55,7 +64,7 @@ class RollModel(db.Model):
             session = db.session
         return(
             session.query(cls).
-            filter_by(user=user).
+            filter_by(user=user["id"]).
             order_by('created desc').
             limit(how_many).
             all()
@@ -67,19 +76,18 @@ class RollModel(db.Model):
         return {
             "id": self.id,
             "key": self.key,
-            "value": self.value,
+            "value": self.val,
             "user": self.user
         }
 
     @property
     def slack_msg(self):
         """Return slack msg."""
-        return "{}: {}".format(self.key, self.value)
+        return "{}: {}".format(self.key, self.val)
 
-    @property
-    def repr(self):
+    def __repr__(self):
         """Repr."""
         return(
             """
             <dungeonbot.models.RollModel(id={},key={}, value={}, user={})>
-            """.format(self.id, self.key, self.value, self.user))
+            """.format(self.id, self.key, self.val, self.user))
